updated code added tracking service added revenue service

+++ src/main/java/com/automation/core/mda/repository/mdaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface mdaRepository extends JpaRepository<mdaModel, Integer> {
+public interface mdaRepository extends JpaRepository<mdaModel, Long> {
     boolean existsByCode(String code);
     void deleteByCode(String code);
     Optional<mdaModel> findBycode(String code);
